chore(plot): rename message_flow_source to message_flow
